TravelManagement/src/main/java/com/crossover/project/entities/User.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package com.crossover.project.entities;

/**
 *
 * @author waqas.ali
 */
public class User {

    private String email;
    private String password;
    private String accountId;
    private String role;
    private String currency;

    public String getCurrency() {
        return currency;
    }

    public void setCurrency(String currency) {
        this.currency = currency;
    }

    public String getAccountId() {
        return accountId;
    }

    public void setAccountId(String accountId) {
        this.accountId = accountId;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
    
    
}
